int -> long :)

+++ lab/src/Main.java
@@ -5,12 +5,12 @@ public class Main {
     static PrintStream out = new PrintStream(System.out);
     static Scanner in = new Scanner(System.in);
 
-    static int [] double_to_int(double number) {
-        int [] answer = new int[2];
+    static long [] double_to_int(double number) {
+        long [] answer = new long[2];
         answer[1] = -1;
         while (number * Math.pow(100, answer[1]) != answer[0] && number * Math.pow(100, answer[1]) != answer[0] * 10) {
             answer[1]++;
-            answer[0] = (int) (number * Math.pow(100, answer[1]));
+            answer[0] = (long) (number * Math.pow(100, answer[1]));
         }
 
         return answer;
@@ -18,9 +18,9 @@ public class Main {
 
 
     static double sqrt(double double_number, int count_accuracy) {
-        int [] int_number = double_to_int(double_number);
-        int number = int_number[0];
-        int divider = int_number[1];
+        long [] int_number = double_to_int(double_number);
+        long number = int_number[0];
+        long divider = int_number[1];
         double accuracy = 1 / Math.pow(10, count_accuracy);
         boolean flag = true;
         double step = number;
